feat: route helper bot generation through pro
