--- packages/service/core/chat/HelperBot/dispatch/index.ts
@@ -1,6 +1,0 @@
-import { HelperBotTypeEnum } from '@fastgpt/global/core/chat/helperBot/type';
-import { dispatchTopAgent } from './topAgent';
-
-export const dispatchMap = {
-  [HelperBotTypeEnum.topAgent]: dispatchTopAgent
-};
