--- packages/service/core/chat/HelperBot/dispatch/topAgent/utils.ts
@@ -1,223 +1,0 @@
-import type { localeType } from '@fastgpt/global/common/i18n/type';
-import { parseI18nString } from '@fastgpt/global/common/i18n/utils';
-import type { ExecutionPlanType, TopAgentGenerationAnswerType } from './type';
-import { SubAppIds, systemSubInfo } from '@fastgpt/global/core/workflow/node/agent/constants';
-import { MongoDataset } from '../../../../dataset/schema';
-import { MongoResourcePermission } from '../../../../../support/permission/schema';
-import { PerResourceTypeEnum } from '@fastgpt/global/support/permission/constant';
-import { getGroupsByTmbId } from '../../../../../support/permission/memberGroup/controllers';
-import { getOrgIdSetWithParentByTmbId } from '../../../../../support/permission/org/controllers';
-import { getUserAvaliableWorkflowTools } from '../../../../app/tool/workflowTool';
-import { SystemToolRepo } from '../../../../app/tool/systemTool/systemTool.repo';
-import { AGENT_SANDBOX_TOOLSET_ID } from '@fastgpt/global/core/ai/sandbox/tools';
-
-const getAccessibleDatasets = async ({ teamId, tmbId }: { teamId: string; tmbId: string }) => {
-  const [roleList, myGroupMap, myOrgSet] = await Promise.all([
-    MongoResourcePermission.find({
-      resourceType: PerResourceTypeEnum.dataset,
-      teamId,
-      resourceId: { $exists: true }
-    }).lean(),
-    getGroupsByTmbId({ tmbId, teamId }),
-    getOrgIdSetWithParentByTmbId({ teamId, tmbId })
-  ]);
-
-  const groupIdSet = new Set(myGroupMap.map((item) => String(item._id)));
-
-  const datasetIds = roleList
-    .filter(
-      (item) =>
-        String(item.tmbId) === String(tmbId) ||
-        (item.groupId && groupIdSet.has(String(item.groupId))) ||
-        (item.orgId && myOrgSet.has(String(item.orgId)))
-    )
-    .map((item) => String(item.resourceId));
-
-  if (datasetIds.length === 0) return [];
-
-  return MongoDataset.find({
-    _id: { $in: Array.from(new Set(datasetIds)) },
-    teamId,
-    deleteTime: null
-  })
-    .select('_id name intro avatar vectorModel')
-    .sort({ updateTime: -1 })
-    .lean();
-};
-
-export const generateResourceList = async ({
-  teamId,
-  tmbId,
-  isRoot,
-  lang = 'zh-CN'
-}: {
-  teamId: string;
-  tmbId: string;
-  isRoot: boolean;
-  lang?: localeType;
-}): Promise<{
-  resourceList: string;
-}> => {
-  const getPrompt = ({ tool, dataset }: { tool: string; dataset: string }) => {
-    return `## 可用工具与知识库
-### 工具
-${tool}
-
-### 知识库
-${dataset}
-
-## 可配置前端开关（不是工具，不能 @ 引用）
-- **file_upload**: 文件上传开关，允许用户在对话中上传文件
-- **sandbox**: 虚拟机开关，允许 Agent 使用虚拟机执行环境
-`;
-  };
-
-  const systemToolRepo = SystemToolRepo.getInstance();
-
-  const [systemTools, myTools, myDatasets] = await Promise.all([
-    systemToolRepo
-      .getSystemToolList({
-        sources: [
-          'system'
-          // teamId
-        ],
-        lang
-      })
-      .then((res) =>
-        res.map((tool) => {
-          const toolId = tool.id;
-          const name = tool.name;
-          const intro = tool.intro;
-          const description = tool.toolDescription || intro || '暂无描述';
-
-          return `- **${toolId}** [工具]: ${name} - ${description}`;
-        })
-      ),
-    getUserAvaliableWorkflowTools({ teamId, tmbId }).then((res) =>
-      res.map((tool) => {
-        const toolId = tool._id;
-        return `- **${toolId}** [工具]: ${tool.name} - ${tool.intro}`;
-      })
-    ),
-    getAccessibleDatasets({ teamId, tmbId }).then((res) => {
-      return res.map((dataset) => {
-        const id = String(dataset._id);
-        const name = dataset.name || '未命名知识库';
-        const intro = dataset.intro || '暂无描述';
-        return `- **${id}** [知识库]: ${name} - ${intro}`;
-      });
-    })
-  ]);
-
-  const builtinTools = [SubAppIds.readFiles, AGENT_SANDBOX_TOOLSET_ID].map((id) => {
-    const info = systemSubInfo[id];
-    return `- **${id}** [工具]: ${parseI18nString(info.name, lang)} - ${info.toolDescription}`;
-  });
-
-  const allTools = [...systemTools, ...myTools, ...builtinTools];
-  return {
-    resourceList: getPrompt({
-      tool: allTools.length > 0 ? allTools.join('\n') : '暂无已安装的工具',
-      dataset: myDatasets.length > 0 ? myDatasets.join('\n') : '暂未配置知识库'
-    })
-  };
-};
-
-/**
- * 从 execution_plan 中提取并去重所有使用的资源
- */
-export const extractResourcesFromPlan = (executionPlan?: ExecutionPlanType) => {
-  if (!executionPlan) {
-    return { tools: [], knowledges: [] };
-  }
-
-  const toolSet = new Set<string>();
-  const knowledgeSet = new Set<string>();
-
-  executionPlan.steps.forEach((step) => {
-    step.expectedTools?.forEach((resourceRef) => {
-      if (resourceRef.type === 'tool') {
-        toolSet.add(resourceRef.id);
-      } else if (resourceRef.type === 'knowledge') {
-        knowledgeSet.add(resourceRef.id);
-      }
-    });
-  });
-
-  return {
-    tools: Array.from(toolSet),
-    knowledges: Array.from(knowledgeSet)
-  };
-};
-
-/**
- * 构建包含所有信息的 system prompt 文本
- * 使用 {{@toolId@}} 格式引用工具，可被 parseSystemPrompt 解析
- */
-export const buildSystemPrompt = (data: TopAgentGenerationAnswerType): string => {
-  const parts: string[] = [];
-
-  // 1. 任务分析
-  if (data.task_analysis) {
-    const { goal, role, key_features } = data.task_analysis;
-    parts.push(`---\n**任务目标**\n${goal}\n`);
-    parts.push(`**角色定位**\n${role}\n`);
-    if (key_features) {
-      parts.push(`**核心特征**\n${key_features}\n`);
-    }
-  }
-
-  // 2. 执行计划
-  if (data.execution_plan) {
-    parts.push(`---\n**参考计划**`);
-    data.execution_plan.steps.forEach((step, index) => {
-      let description = step.description;
-
-      // 替换 description 中的资源引用：
-      // - 工具: @工具ID / @工具ID@ / @[工具ID] -> {{@工具ID@}}
-      // - 知识库: @知识库ID / @知识库ID@ / @[知识库ID] -> {{@dataset_search@}}
-      if (step.expectedTools && step.expectedTools.length > 0) {
-        step.expectedTools.forEach((resourceRef) => {
-          const replaceId =
-            resourceRef.type === 'knowledge' ? SubAppIds.datasetSearch : resourceRef.id;
-          const escapedId = resourceRef.id.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-          const regex = new RegExp(`@(?:\\[${escapedId}\\]|${escapedId}@?)`, 'g');
-          description = description.replace(regex, `{{@${replaceId}@}}`);
-        });
-      }
-      description = description.replace(
-        /(?<!\{\{)@(?:\[(file_upload|sandbox)\]|(file_upload|sandbox)@?)(?!\}\})/g,
-        '$1$2'
-      );
-
-      parts.push(`\n步骤 ${index + 1}. ${step.title} \n${description}`);
-      // if (step.expectedTools && step.expectedTools.length > 0) {
-      //   const toolList = step.expectedTools
-      //     .map((t) => {
-      //       const ref = `{{@${t.id}@}}`;
-      //       return `${ref}`;
-      //     })
-      //     .join('、');
-      //   parts.push(`预期资源: ${toolList}`);
-      // }
-    });
-    parts.push('');
-  }
-
-  // 3. 系统功能
-  // if (data.resources?.system_features?.file_upload?.enabled) {
-  //   parts.push(`---\n**系统功能**\n`);
-  //   parts.push(
-  //     `**文件上传**: 已启用\n${data.resources.system_features.file_upload.purpose}`
-  //   );
-  // }
-
-  return parts.join('\n');
-};
-
-/**
- * 构建用于显示的文本（与 system prompt 格式一致）
- */
-export const buildDisplayText = (data: TopAgentGenerationAnswerType): string => {
-  return buildSystemPrompt(data);
-};
