+++ projects/app/src/components/core/chat/ChatContainer/type.ts
@@ -11,7 +11,7 @@ import type {
   UserInputInteractive,
   WorkflowInteractiveResponseType
 } from '@fastgpt/global/core/workflow/template/system/interactive/type';
-import type { TopAgentFormDataType } from '@fastgpt/service/core/chat/HelperBot/dispatch/topAgent/type';
+import type { TopAgentFormDataType } from '@fastgpt/global/core/chat/helperBot/topAgent/type';
 import type { AgentPlanStatusType, AgentPlanType } from '@fastgpt/global/core/ai/agent/type';
 
 export type generatingMessageProps = {
